# swagger.yaml
openapi: 3.0.0
info:
  title: Warehouse Co
  description: Warehouse and Inventory management system restful Api
  contact: {}
  version: "1.0"
servers:
  - url: https://odd-rose-turkey-wear.cyclic.app/api/v1
    variables: {}
paths:
  /products:
    post:
      tags:
        - Product
      summary: Create Product
      description: Create new product for the admin or owner
      operationId: CreateProduct
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - product
                - price
                - name
                - category
              type: object
              properties:
                product:
                  type: string
                price:
                  type: integer
                  format: int32
                  example: 32000
                name:
                  type: string
                  example: Bag of rice
                category:
                  type: string
                  example: Food
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    get:
      tags:
        - Product
      summary: Show all Product
      description: Show all products for all
      operationId: ShowallProduct
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /product/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The product ID
    delete:
      tags:
        - Product
      summary: Delete Product
      description: Delete Product for the owner
      operationId: DeleteProduct
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false

    get:
      tags:
        - Product
      summary: Show Single Product
      description: Show single product
      operationId: ShowSingleProduct
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    patch:
      tags:
        - Product
      summary: Edit Product
      description: Make edits to product for owner
      operationId: EditProduct
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - price
              type: object
              properties:
                price:
                  type: integer
                  format: int32
                  example: 45000
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
components:
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...